mojang more deep removal
