+++ build.gradle
@@ -46,10 +46,8 @@ dependencies {
     provided fileTree(dir: 'src/main/libraries', include: '*.jar')
     provided 'com.squareup.okhttp3:okhttp-tls:4.9.2'
     provided 'com.squareup.okhttp3:okhttp:4.9.2'
-    //provided 'com.mojang:authlib:1.5.21'
     provided 'com.google.code.gson:gson:2.8.5'
     provided 'com.google.http-client:google-http-client:1.23.0'
-    provided 'com.github.Cydhra:Nidhogg:1.3'
     provided 'com.google.guava:guava:28.0-jre'
     provided 'org.apache.commons:commons-lang3:3.9'
     provided 'commons-io:commons-io:2.6'
