+++ src/main/java/net/dirtcraft/dirtlauncher/game/LaunchGame.java
@@ -13,7 +13,7 @@ import net.dirtcraft.dirtlauncher.data.DirtLauncher.Settings;
 import net.dirtcraft.dirtlauncher.data.Minecraft.JavaVersion;
 import net.dirtcraft.dirtlauncher.exceptions.InstanceException;
 import net.dirtcraft.dirtlauncher.exceptions.LaunchException;
-import net.dirtcraft.dirtlauncher.game.authentification.account.Account;
+import net.dirtcraft.dirtlauncher.game.authentification.Account;
 import net.dirtcraft.dirtlauncher.game.modpacks.Modpack;
 import net.dirtcraft.dirtlauncher.game.serverlist.Listing;
 import net.dirtcraft.dirtlauncher.game.serverlist.ServerList;
