+++ src/main/java/net/dirtcraft/dirtlauncher/game/authentification/LoginError.java
@@ -1,6 +1,5 @@
 package net.dirtcraft.dirtlauncher.game.authentification;
 
-import net.cydhra.nidhogg.exception.*;
 import net.dirtcraft.dirtlauncher.gui.dialog.ErrorWindow;
 
 public enum LoginError {
@@ -8,12 +7,6 @@ public enum LoginError {
     INVALID_CREDENTIALS, INVALID_SESSION, TOO_MANY_REQUESTS, UNAUTHORISED_OPERATION, USER_MIGRATED, YGGDRASIL_BAN, ILLEGAL_ARGUMENT, UNKNOWN;
 
     public static LoginError from(Exception yggdrasilClientException){
-        if (yggdrasilClientException instanceof InvalidCredentialsException) return INVALID_CREDENTIALS;
-        if (yggdrasilClientException instanceof InvalidSessionException) return INVALID_SESSION;
-        if (yggdrasilClientException instanceof TooManyRequestsException) return TOO_MANY_REQUESTS;
-        if (yggdrasilClientException instanceof UnauthorizedOperationException) return UNAUTHORISED_OPERATION;
-        if (yggdrasilClientException instanceof UserMigratedException) return USER_MIGRATED;
-        if (yggdrasilClientException instanceof YggdrasilBanException) return YGGDRASIL_BAN;
         if (yggdrasilClientException instanceof IllegalArgumentException) {
             new ErrorWindow(yggdrasilClientException.getLocalizedMessage());
             return ILLEGAL_ARGUMENT;
